backend/requirements.txt
Django==4.2.11
django-cors-headers==4.3.1
Faker==24.0.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.15.2
psycopg-binary==3.1.18
psycopg==3.1.18
gunicorn==22.0.0
beautifulsoup4==4.12.3
requests==2.32.2
lxml==5.1.0
rsa==4.9
cryptography==42.0.5
boto3==1.34.75
django-storages==1.14.2
botocore==1.34.75
s3transfer==0.10.1
pillow==10.3.0
django-redis==5.4.0
pywebpush==2.0.0
django-rest-knox==4.2.0
pytz==2024.1
